now it's functional
it MAY not be perfect. in fact I know it isn't. I'd like to clean it up
and make it less of a Script and more of a Piece Of Software, and
probably with js instead tbh. there's also a weird bug that happens if
you cancel it just after getting the last month in a year, where it
starts at month 2 of the year instead of month 1 of the next. now that I
write that out, I know why it's happening, but I need to figure out how
to fix it. later. or maybe not. who really cares if I'm being honest